requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain>=0.1.0
langchain-community>=0.0.13
openai>=1.6.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0
playwright>=1.40.0